+++ apps/docs/src/lib/components.ts
@@ -888,8 +888,9 @@ export function getComponents(): ComponentDetail[] {
         {
           platform: {
             id: 'react',
+            href: '?path=/docs/navigation-pagination--docs',
           },
-          status: 'considering',
+          status: 'beta',
         },
       ],
     },