+++ packages/html/ds/src/index.ts
@@ -4,6 +4,7 @@ import { createEvents } from './events.js';
 import { initHeader } from './header/header.js';
 import { initCheckboxes } from './checkbox/checkbox.js'; /* eslint-disable-line import/order */
 import { initComboBox } from './combo-box/combo-box.js'; /* eslint-disable-line import/order */
+import { initAlert } from './alert/alert.js'; /* eslint-disable-line import/order */
 import { initModal } from './modal/modal.js';
 import { initRadios } from './radio/radio.js';
 import { initTabs } from './tabs/tabs.js';
@@ -34,6 +35,7 @@ export function initGovIe() {
   initCookieBanner();
   initTabs();
   initComboBox();
+  initAlert();
 }
 
 export function destroyGovIe() {