+++ packages/react/ds/src/container/container.tsx
@@ -1,6 +1,12 @@
-export function Container({ children }: { children: React.ReactNode }) {
+export function Container({
+  children,
+  id,
+}: {
+  children: React.ReactNode;
+  id?: string;
+}) {
   return (
-    <div data-testid="govie-container" className="gi-layout-container">
+    <div data-testid="govie-container" className="gi-layout-container" id={id}>
       {children}
     </div>
   );