+++ packages/react/ds/src/radio/radio.stories.tsx
@@ -1,241 +1,92 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { RadiosGroup } from './radios-group.js';
-import type { RadiosGroupType } from './types.js';
-
-const standardProps: RadiosGroupType = {
-  fieldId: 'uniqueId',
-  title: {
-    value: 'Where do you live?',
-    asHeading: {
-      size: 'md',
-      as: 'h2',
-    },
-  },
-  items: [
-    {
-      label: 'England',
-      value: 'england',
-    },
-    {
-      label: 'Scotland',
-      value: 'scotland',
-    },
-    {
-      label: 'Ireland',
-      value: 'ireland',
-    },
-  ],
-};
+import { Radio } from './radio.js';
 
 const meta = {
-  title: 'Form/Radios',
+  title: 'Form/Radio/Radio',
   parameters: {
     docs: {
       description: {
-        component:
-          'Radio component when users can only select one option from a list.',
+        component: 'Radio component',
       },
     },
   },
-  component: RadiosGroup,
-} satisfies Meta<typeof RadiosGroup>;
+  component: Radio,
+} satisfies Meta<typeof Radio>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
   argTypes: {
-    fieldId: {
+    value: {
       control: 'text',
-      type: { name: 'string', required: true },
-      description: 'The unique value for the radios group',
+      type: 'string',
+      description: 'The value of the Component',
     },
-    items: {
-      control: 'object',
-      type: {
-        name: 'object',
-        required: true,
-        value: {
-          label: {
-            name: 'string',
-          },
-          value: {
-            name: 'string',
-            required: true,
-          },
-          hint: {
-            name: 'string',
-          },
-          conditionalInput: {
-            name: 'other',
-            value:
-              'Conditional input associated with the radio (see text input component)',
-          },
-        },
-      },
+    name: {
+      control: 'text',
+      type: 'string',
+      description: 'Name attribute of the input element',
+    },
+    label: {
+      control: 'text',
+      type: 'string',
+      description: 'The label associated with the Radio',
+    },
+    hint: {
+      control: 'text',
+      type: 'string',
       description:
-        'A list of the label, values, hint and conditional for the radios',
+        'Additional text to inform the user about the Radio component',
     },
-    inline: {
-      control: 'boolean',
-      description: 'Specify if the radios should be inline',
-    },
-    errorMessage: {
+    id: {
       control: 'text',
-      type: { name: 'string' },
-      description: 'The text of the error message',
-    },
-    title: {
-      control: 'object',
-      type: {
-        name: 'object',
-        value: {
-          value: {
-            name: 'string',
-          },
-          asHeading: {
-            name: 'boolean',
-          },
-          hint: {
-            name: 'string',
-          },
-        },
-      },
-      description: 'The label, hint and option the make the title a heading',
+      type: 'string',
+      description: 'The id of the Radio',
     },
     size: {
       control: 'radio',
       options: ['lg', 'md', 'sm'],
-      description: 'The sizes of the checkboxes',
+      description: 'The sizes for the Radio',
+    },
+    conditionalInput: {
+      description:
+        'Conditional input associated with the radio (see text input component)',
+    },
+    checked: {
+      control: 'boolean',
+      description: 'if true the component is checked',
+    },
+    onChange: {
+      control: 'object',
+      type: 'function',
+      description: 'Callback fired when the radio component is selected.',
     },
   },
   args: {
-    ...standardProps,
-    fieldId: 'UniqueId1',
+    value: 'radio-value',
+    label: 'Radio',
   },
 };
 
-export const inline: Story = {
+export const withHint: Story = {
   args: {
-    ...standardProps,
-    fieldId: 'UniqueId2',
-    inline: true,
+    value: 'radio-with-hint',
+    label: 'With hint',
+    hint: 'This is a hint',
   },
 };
 
-export const withTitleHint: Story = {
+export const withDefaultChecked: Story = {
   args: {
-    ...standardProps,
-    title: {
-      value: 'Where do you live?',
-      hint: 'Specify the location where you live',
-      asHeading: {
-        size: 'md',
-        as: 'h2',
-      },
-    },
-    fieldId: 'UniqueId3',
+    value: 'radio-with-default-checked',
+    label: 'Default checked',
+    checked: true,
   },
 };
 
-export const withOptionHints: Story = {
+export const withoutLabel: Story = {
   args: {
-    title: {
-      value: 'Have you changed your name?',
-      hint: 'This includes changing your last name or spelling your name differently.',
-      asHeading: {
-        size: 'md',
-        as: 'h2',
-      },
-    },
-    items: [
-      {
-        label: 'Yes',
-        value: 'yes',
-        hint: 'Yes, I have changed my name',
-      },
-      {
-        label: 'No',
-        value: 'no',
-        hint: "No, I didn't change my name",
-      },
-    ],
-    fieldId: 'UniqueId4',
-  },
-};
-
-export const withDividerOption: Story = {
-  args: {
-    ...standardProps,
-    dividerOption: {
-      label: 'None of the above',
-      value: 'none of the above',
-    },
-    fieldId: 'UniqueId5',
-  },
-};
-
-export const withError: Story = {
-  args: {
-    ...standardProps,
-    errorMessage: 'Please select an option',
-    fieldId: 'UniqueId6',
-  },
-};
-
-export const withConditionalInput: Story = {
-  args: {
-    title: {
-      value: 'How would you prefer to be contacted?',
-      asHeading: {
-        size: 'md',
-        as: 'h2',
-      },
-    },
-    items: [
-      {
-        label: 'Email',
-        value: 'email',
-        conditionalInput: {
-          id: 'input-id-email',
-          label: {
-            text: 'Email address',
-          },
-          type: 'email',
-        },
-      },
-      {
-        label: 'Phone',
-        value: 'phone',
-        conditionalInput: {
-          id: 'input-id-phone',
-          label: {
-            text: 'Phone number',
-          },
-          type: 'tel',
-        },
-      },
-    ],
-    fieldId: 'UniqueId7',
-  },
-};
-
-export const withNoTitle: Story = {
-  args: {
-    fieldId: 'uniqueId8',
-    items: [
-      {
-        label: 'England',
-        value: 'england',
-      },
-      {
-        label: 'Scotland',
-        value: 'scotland',
-      },
-      {
-        label: 'Ireland',
-        value: 'ireland',
-      },
-    ],
+    value: 'radio-without-label',
   },
 };